CFPlanets

Documentation for CFPlanets.

CFPlanets.ConformalPlanet

CFPlanets.ShallowTradPlanet

CFPlanets.coriolis

CFPlanets.scale_factor

CFPlanets.ConformalPlanet — Type

abstract type ConformalPlanet end

An abstract type for planets such that the two metric factors are equal.

CFPlanets.ShallowTradPlanet — Type

struct ShallowTradPlanet{F} <: ConformalPlanet
planet = ShallowTradPlanet(radius, Omega)

Instances of this type describe a spherical-geoid, shallow-atmosphere, traditional approximation. radius is the radius of the planet and Omega its rotation rate.

CFPlanets.coriolis — Function

f = coriolis(planet, lon, lat) # spherical geometry
f = coriolis(planet, x, y)     # logically Cartesian

Returns the Coriolis parameter corresponding to geometric approximation planet.

CFPlanets.scale_factor — Function

factor = scale_factor(planet::ConformalPlanet, lon, lat) # spherical domain
factor = scale_factor(planet::ConformalPlanet, x, y)     # logically Cartesian domain

Returns the scale factor at a certain point on planet`, given by its abstract coordinates. The scale factor is related to the physical length of a horizontal displacement by:

δl = scale_factor * sqrt(δlat^2 + cos(lat)^2 δlon^2) # or
δl = scale_factor * sqrt(δx^2 + δy^2).

For instance if the the spherical-geoid and shallow-atmosphere approximations are made, the scale factor is the planetary radius, assumed constant.